move header and footer into navigation
