+++ src/routes/register/+page.svelte
@@ -2,11 +2,10 @@
 	import type { LoadRegister } from 'types/types';
 	export let data: LoadRegister; // data from database
 
-	import Header from 'elements/header.svelte';
+	import Header from 'elements/navigation/header.svelte';
 	import HeadlinePage from 'elements/text/headlinePage.svelte';
-	import Footer from 'elements/footer.svelte';
+	import Footer from 'elements/navigation/footer.svelte';
 
-	import HorizontalLine from 'elements/line/horizontalLine.svelte';
 	import Spacer from 'elements/spacer/spacer.svelte';
 	import MessageWrapper from 'elements/messages/messageWrapper.svelte';
 	import ErrorMessage from 'elements/messages/errorMessage.svelte';
